pyproject.toml
[project]
name = "telegram-news-bot"
version = "1.0.0"
description = "Telegram bot that aggregates news headlines from multiple sources"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic-settings>=2.0",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "ruff>=0.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
